refactor: shared Jest setup to eliminate duplication
Extract common Jest setup into packages/shared/src/jest.setup.base.ts
Reduces ~31 lines of duplicated setup code
Packages now re-export from shared location
Maintains 100% backward compatibility

Closes #482
